Add security tests for various controllers and implement user service tests Created security tests for ClientVisitController, ListingController, NotificationController, SavedSearchController to ensure proper authentication and authorization. Implemented unit tests for UserController to validate user creation for agents and managers, including email notifications. Added unit tests for JwtService to verify token generation and validation.

+++ Backend-DietiEstates25/src/main/java/it/unina/dietiestates25/backend/services/DashboardStatsService.java
@@ -39,7 +39,7 @@ public class DashboardStatsService {
 
     public ClientStatsResponse getClientStats(UUID clientId) {
         long totalVisits = visitRepository.findAllByClient_Id(clientId).size();
-        long pendingVisits = visitRepository.findAllByClient_IdAndStatus(clientId, VisitStatus.REQUESTED).size() +
+        long pendingVisits = (long) visitRepository.findAllByClient_IdAndStatus(clientId, VisitStatus.REQUESTED).size() +
                             visitRepository.findAllByClient_IdAndStatus(clientId, VisitStatus.CONFIRMED).size();
         long completedVisits = visitRepository.findAllByClient_IdAndStatus(clientId, VisitStatus.DONE).size();
